Move the DisplayWork-publishers serialisation tests to the display lib
