+++ sbt_common/display/src/test/scala/uk/ac/wellcome/display/models/DisplayWorkSerialisationTest.scala
@@ -528,4 +528,45 @@ class DisplayWorkSerialisationTest
       assertJsonStringsAreEqual(actualJson, expectedJson)
     }
   }
+
+  describe("publishers") {
+    it(
+      "includes the publishers field with a mixture of agents/organisations/persons") {
+      val work = IdentifiedWork(
+        canonicalId = "v9w6cz66",
+        sourceIdentifier = sourceIdentifier,
+        version = 1,
+        title = Some("Vultures vying for victory"),
+        publishers = List(
+          Unidentifiable(Agent("Vivian Violet")),
+          Unidentifiable(Organisation("Verily Volumes")),
+          Unidentifiable(
+            Person(
+              label = "Havelock Vetinari",
+              prefix = Some("Lord Patrician"),
+              numeration = Some("I")))
+        )
+      )
+      val displayWork = DisplayWork(work)
+
+      val actualJson = objectMapper.writeValueAsString(displayWork)
+      val expectedJson = s"""
+                            |{
+                            |  "type": "Work",
+                            |  "id": "${work.canonicalId}",
+                            |  "title": "${work.title.get}",
+                            |  "creators": [ ],
+                            |  "subjects": [ ],
+                            |  "genres": [ ],
+                            |  "publishers": [
+                            |    ${identifiedOrUnidentifiable(work.publishers(0), abstractAgent)},
+                            |    ${identifiedOrUnidentifiable(work.publishers(1), abstractAgent)},
+                            |    ${identifiedOrUnidentifiable(work.publishers(2), abstractAgent)}
+                            |  ],
+                            |  "placesOfPublication": [ ]
+                            |}""".stripMargin
+
+      assertJsonStringsAreEqual(actualJson, expectedJson)
+    }
+  }
 }
